react-django-ecommerce-main/rde-backend/requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.1
colorama==0.4.4
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
